Confirm to ofxAddonsTemplate + Modernization.
