+++ example-map/openFrameworks-Info.plist
