--- example/ofxTweenExample/src/main.cpp
@@ -1,16 +1,0 @@
-#include "ofMain.h"
-#include "testApp.h"
-#include "ofAppGlutWindow.h"
-
-//========================================================================
-int main( ){
-
-    ofAppGlutWindow window;
-	ofSetupOpenGL(&window, 1280,800, OF_WINDOW);			// <-------- setup the GL context
-
-	// this kicks off the running of my app
-	// can be OF_WINDOW or OF_FULLSCREEN
-	// pass in width and height too:
-	ofRunApp( new testApp());
-
-}
